Fixed bug where depositing the last Pokemon in your party would crash the game
Fixed a major bug where depositing the Pokemon in the last slot of your party would cause the game to crash. This was due to code I had written to zero out happiness data after shifting back, but this code only worked correctly if you were removing a Pokemon from the beginning or middle of the party. It now works as intended, even if it is a bit clunky.